Refactor genesis and chain verification
